Docker App - Flask with MySQL/server/requirements.txt
Flask
mysql-connector-python
python-dotenv
gunicorn
uvicorn
waitress